@@ -1,6 +1,6 @@
 {
   "name": "bento_web",
-  "version": "4.1.4",
+  "version": "5.0.0",
   "description": "Bento platform front-end",
   "main": "src/index.js",
   "dependencies": {
@@ -9,46 +9,41 @@
     "ajv": "^8.12.0",
     "ansi_up": "^6.0.2",
     "antd": "^4.24.15",
-    "bento-auth-js": "^3.2.6",
-    "bento-charts": "~2.3.0",
-    "cross-fetch": "^3.1.8",
+    "bento-auth-js": "^4.1.0",
+    "bento-charts": "~2.6.8",
+    "cross-fetch": "^4.0.0",
     "file-saver": "^2.0.5",
     "filesize": "^10.0.12",
     "handlebars": "^4.7.8",
     "igv": "~2.15.11",
     "iso8601-duration": "^1.2.0",
-    "jose": "^4.14.6",
     "lodash": "^4.17.21",
     "lodash.debounce": "^4.0.8",
     "mammoth": "^1.6.0",
     "papaparse": "^5.4.1",
     "prop-types": "^15.7.2",
     "query-string": "^6.13.7",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.3",
     "react-icons": "^4.11.0",
     "react-markdown": "^8.0.7",
     "react-pdf": "^7.7.0",
-    "react-redux": "^7.2.9",
-    "react-router-dom": "^5.3.4",
-    "react-router-prop-types": "^1.0.5",
+    "react-redux": "^8.1.3",
+    "react-router-dom": "^6.22.3",
     "react-syntax-highlighter": "^13.5.3",
     "react18-json-view": "^0.2.7",
-    "redux": "^4.2.0",
+    "redux": "^4.2.1",
     "redux-thunk": "^2.4.2",
-    "remark-gfm": "^3.0.1",
+    "remark-gfm": "^4.0.0",
     "socket.io-client": "^4.7.2",
-    "victory": "^35.3.1",
     "xlsx": "https://cdn.sheetjs.com/xlsx-0.20.0/xlsx-0.20.0.tgz"
   },
   "devDependencies": {
-    "@babel/core": "^7.23.0",
-    "@babel/eslint-parser": "^7.22.15",
-    "@babel/plugin-proposal-class-properties": "^7.18.6",
-    "@babel/plugin-proposal-optional-chaining": "^7.20.7",
-    "@babel/preset-env": "^7.22.20",
-    "@babel/preset-react": "^7.22.15",
+    "@babel/core": "^7.24.1",
+    "@babel/eslint-parser": "^7.24.1",
+    "@babel/preset-env": "^7.24.1",
+    "@babel/preset-react": "^7.24.1",
     "babel-loader": "^9.1.3",
     "babel-plugin-import": "^1.13.8",
     "babel-polyfill": "^6.26.0",
@@ -62,7 +57,7 @@
     "style-loader": "^3.3.1",
     "webpack": "^5.88.2",
     "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^4.15.1",
+    "webpack-dev-server": "^5.0.3",
     "worker-loader": "^3.0.8"
   },
   "scripts": {
