@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { useState, useEffect, useCallback } from "react";
 import { useSelector } from "react-redux";
 import PropTypes from "prop-types";
 
@@ -8,22 +8,37 @@ import LocusSearch from "./LocusSearch";
 
 import { notAlleleCharactersRegex } from "@/utils/misc";
 
-const VariantSearchHeader = ({dataType, addVariantSearchValues}) => {
-    const [refFormReceivedValidKeystroke, setRefFormReceivedValidKeystroke ] = useState(true);
-    const [altFormReceivedValidKeystroke , setAltFormReceivedValidKeystroke ] = useState(true);
+
+const isValidLocus = (locus) =>
+    locus.chrom !== null && locus.start !== null && locus.end !== null;
+const normalizeAlleleText = (text) =>
+    text.toUpperCase().replaceAll(notAlleleCharactersRegex, "");
+const containsInvalid = (text) => {
+    const matches = text.toUpperCase().match(notAlleleCharactersRegex);
+    return matches && matches.length > 0;
+};
+
+const INITIAL_FIELDS_VALIDITY = {
+    "assemblyId": true,
+    "locus": true,
+};
+
+// Match style from DiscoverySearchForm
+const LABEL_COL = { lg: { span: 24 }, xl: { span: 4 }, xxl: { span: 3 } };
+const WRAPPER_COL = { lg: { span: 24 }, xl: { span: 20 }, xxl: { span: 18 } };
+
+
+const VariantSearchHeader = ({ dataType, addVariantSearchValues }) => {
+    const [refFormReceivedValidKeystroke, setRefFormReceivedValidKeystroke] = useState(true);
+    const [altFormReceivedValidKeystroke, setAltFormReceivedValidKeystroke] = useState(true);
     const [activeRefValue, setActiveRefValue] = useState(null);
     const [activeAltValue, setActiveAltValue] = useState(null);
     const [assemblyId, setAssemblyId] = useState(null);
-    const [locus, setLocus] = useState({chrom: null, start: null, end: null});
+    const [locus, setLocus] = useState({ chrom: null, start: null, end: null });
     const isSubmitting = useSelector(state => state.explorer.isSubmittingSearch);
 
     // begin with required fields considered valid, so user isn't assaulted with error messages
-    const initialValidity = {
-        "assemblyId": true,
-        "locus": true,
-    };
-
-    const [fieldsValidity, setFieldsValidity] = useState(initialValidity);
+    const [fieldsValidity, setFieldsValidity] = useState(INITIAL_FIELDS_VALIDITY);
 
     const variantsOverviewResults = useSelector((state) => state.explorer.variantsOverviewResponse);
     const hasAssemblyIds =
@@ -34,10 +49,6 @@ const VariantSearchHeader = ({dataType, addVariantSearchValues}) => {
     const assemblySchema = dataType.schema?.properties?.assembly_id;
     const genotypeSchema = dataType.schema?.properties?.calls?.items?.properties?.genotype_type;
 
-  // hardcoded style from DiscoverySearchForm, change to params
-    const labelCol = {lg: { span: 24 }, xl: { span: 4 }, xxl: { span: 3 }};
-    const wrapperCol = {lg: { span: 24 }, xl: { span: 20 }, xxl: { span: 18 }};
-
     const helpText = {
         "assemblyId": assemblySchema?.description,
         "genotype": genotypeSchema?.description,
@@ -46,23 +57,22 @@ const VariantSearchHeader = ({dataType, addVariantSearchValues}) => {
     };
 
     // custom validation since this form isn't submitted, it's just used to fill fields in hidden form
-    // each field is validated invididually elsewhere
+    // each field is validated individually elsewhere
     // for final validation, we only need to make sure required fields are non-empty
-    const validateVariantSearchForm = () => {
-
-      // check assembly
+    const validateVariantSearchForm = useCallback(() => {
+        // check assembly
         if (!assemblyId) {
-        // change assemblyId helptext & outline
-            setFieldsValidity({...fieldsValidity, "assemblyId": false});
+            // change assemblyId help text & outline
+            setFieldsValidity({ ...fieldsValidity, "assemblyId": false });
         }
 
         // check locus
-        const {chrom, start, end} = locus;
+        const { chrom, start, end } = locus;
         if (!chrom || !start || !end) {
-        // change locus helptext & outline
-            setFieldsValidity({...fieldsValidity, "locus": false});
+            // change locus help text & outline
+            setFieldsValidity({ ...fieldsValidity, "locus": false });
         }
-    };
+    }, [assemblyId, locus, fieldsValidity]);
 
     useEffect(() => {
         if (isSubmitting) {
@@ -70,69 +80,56 @@ const VariantSearchHeader = ({dataType, addVariantSearchValues}) => {
         }
     }, [isSubmitting]);
 
+    const setLocusValidity = useCallback((isValid) => {
+        setFieldsValidity({ ...fieldsValidity, "locus": isValid });
+    }, [fieldsValidity]);
 
-    const isValidLocus = (locus) => {
-        return locus.chrom !== null  && locus.start !== null && locus.end !== null;
-    };
-
-    const setLocusValidity = (isValid) => {
-        setFieldsValidity({...fieldsValidity, "locus": isValid});
-    };
-
-    const handleLocusChange = (locus) => {
+    const handleLocusChange = useCallback((locus) => {
         setLocusValidity(isValidLocus(locus));
 
-        // set even if invalid so we don't keep old values
+        // set even if invalid, so we don't keep old values
         setLocus(locus);
-    };
+    }, [setLocusValidity]);
 
-    const handleAssemblyIdChange = (value) => {
-        addVariantSearchValues({assemblyId: value});
+    const handleAssemblyIdChange = useCallback((value) => {
+        addVariantSearchValues({ assemblyId: value });
         setAssemblyId(value);
-    };
+    }, []);
 
-    const handleGenotypeChange = (value) => {
-        addVariantSearchValues({genotypeType: value});
-    };
+    const handleGenotypeChange = useCallback((value) => {
+        addVariantSearchValues({ genotypeType: value });
+    }, []);
 
-    const handleRefChange = (e) => {
+    const handleRefChange = useCallback((e) => {
         const latestInputValue = e.target.value;
-        const validatedRef = validateAlleleText(latestInputValue);
+        const normalizedRef = normalizeAlleleText(latestInputValue);
         const didValueContainInvalidChars = containsInvalid(latestInputValue);
 
         if (didValueContainInvalidChars) {
             setRefFormReceivedValidKeystroke(!didValueContainInvalidChars);
-            setTimeout(function() {
+            setTimeout(() => {
                 setRefFormReceivedValidKeystroke(true);
             }, 1000);
         }
-        setActiveRefValue(validatedRef);
-        addVariantSearchValues({ref: validatedRef});
-    };
+        setActiveRefValue(normalizedRef);
+        addVariantSearchValues({ ref: normalizedRef });
+    }, []);
 
-    const handleAltChange = (e) => {
+    const handleAltChange = useCallback((e) => {
         const latestInputValue = e.target.value;
-        const validatedAlt = validateAlleleText(latestInputValue);
+        const normalizedAlt = normalizeAlleleText(latestInputValue);
         const didValueContainInvalidChars = containsInvalid(latestInputValue);
 
         if (didValueContainInvalidChars) {
             setAltFormReceivedValidKeystroke(!didValueContainInvalidChars);
-            setTimeout(function() {
+            setTimeout(() => {
                 setAltFormReceivedValidKeystroke(true);
             }, 1000);
         }
 
-        setActiveAltValue(validatedAlt);
-        addVariantSearchValues({alt: validatedAlt});
-    };
-
-    const validateAlleleText = (text) => {
-        return text.toUpperCase().replaceAll(notAlleleCharactersRegex, "");
-    };
-    const containsInvalid = (text) => {
-        const matches = text.toUpperCase().match(notAlleleCharactersRegex);
-        return matches && matches.length > 0;
-    };
+        setActiveAltValue(normalizedAlt);
+        addVariantSearchValues({ alt: normalizedAlt });
+    }, []);
 
     // set default selected assemblyId if only 1 is present
     const shouldTriggerAssemblyIdChange = overviewAssemblyIds.length === 1;
@@ -152,9 +149,9 @@ const VariantSearchHeader = ({dataType, addVariantSearchValues}) => {
     return (
         <>
             <Form.Item
-                labelCol={labelCol}
-                wrapperCol={wrapperCol}
-                label={"Assembly ID"}
+                labelCol={LABEL_COL}
+                wrapperCol={WRAPPER_COL}
+                label="Assembly ID"
                 help={helpText["assemblyId"]}
                 validateStatus={fieldsValidity.assemblyId ? "success" : "error"}
                 required
@@ -162,14 +159,13 @@ const VariantSearchHeader = ({dataType, addVariantSearchValues}) => {
                 <Select
                     onChange={handleAssemblyIdChange}
                     defaultValue={overviewAssemblyIds && shouldTriggerAssemblyIdChange && overviewAssemblyIds[0]}
-                >
-                    {overviewAssemblyIds.map(v => <Select.Option key={v} value={v}>{v}</Select.Option>)}
-                </Select>
+                    options={overviewAssemblyIds.map((value) => ({ value, label: value }))}
+                />
             </Form.Item>
             <Form.Item
-                labelCol={labelCol}
-                wrapperCol={wrapperCol}
-                label={"Gene / position"}
+                labelCol={LABEL_COL}
+                wrapperCol={WRAPPER_COL}
+                label="Gene / position"
                 help={helpText["locus"]}
                 validateStatus={fieldsValidity.locus ? "success" : "error"}
                 required
@@ -180,9 +176,9 @@ const VariantSearchHeader = ({dataType, addVariantSearchValues}) => {
                              setLocusValidity={setLocusValidity} />
             </Form.Item>
             <Form.Item
-                labelCol={labelCol}
-                wrapperCol={wrapperCol}
-                label={"Genotype"}
+                labelCol={LABEL_COL}
+                wrapperCol={WRAPPER_COL}
+                label="Genotype"
                 help={helpText["genotype"]}
             >
                 <Select
@@ -193,9 +189,9 @@ const VariantSearchHeader = ({dataType, addVariantSearchValues}) => {
                 </Select>
             </Form.Item>
             <Form.Item
-                labelCol={labelCol}
-                wrapperCol={wrapperCol}
-                label={"Reference Allele"}
+                labelCol={LABEL_COL}
+                wrapperCol={WRAPPER_COL}
+                label="Reference Allele"
                 help={helpText["ref/alt"]}
             >
                 <Input
@@ -205,9 +201,9 @@ const VariantSearchHeader = ({dataType, addVariantSearchValues}) => {
                 />
             </Form.Item>
             <Form.Item
-                labelCol={labelCol}
-                wrapperCol={wrapperCol}
-                label={"Alternate Allele"}
+                labelCol={LABEL_COL}
+                wrapperCol={WRAPPER_COL}
+                label="Alternate Allele"
                 help={helpText["ref/alt"]}
             >
                 <Input
